requirements.txt
load_dotenv
pymupdf
aiohttp
huggingface_hub
google-genai
transformers